+++ src/cylinder_ops.c
@@ -6,7 +6,7 @@
 /*   By: wchoe <wchoe@student.42gyeongsan.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/30 02:56:47 by chakim            #+#    #+#             */
-/*   Updated: 2025/07/08 16:22:16 by wchoe            ###   ########.fr       */
+/*   Updated: 2025/07/08 16:36:38 by wchoe            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 
 static t_vec4	cylinder_get_normal(t_vec4 p_local)
 {
-	if (fabs(p_local.v[2] - 0.5f) < EPSILON)
+	if (fabsf(p_local.v[2] - 0.5f) < EPSILON)
 		return ((t_vec4){{0.0f, 0.0f, 1.0f, 0.0f}});
-	else if (fabs(p_local.v[2] + 0.5f) < EPSILON)
+	else if (fabsf(p_local.v[2] + 0.5f) < EPSILON)
 		return ((t_vec4){{0.0f, 0.0f, -1.0f, 0.0f}});
 	else
 		return ((t_vec4){{p_local.v[0], p_local.v[1], 0.0f, 0.0f}});
@@ -47,7 +47,7 @@ int	cylinder_intersect(const t_object *obj, \
 	record->n_world = mat_mul_vec4(&obj->n, \
 		cylinder_get_normal(record->p_local));
 	record->n_world = vec4_mul(record->n_world, 1.0f / \
-		sqrt(vec4_dot(record->n_world, record->n_world)));
+		sqrtf(vec4_dot(record->n_world, record->n_world)));
 	if (vec4_dot(record->n_world, ray_world->d) > 0.0f)
 		record->n_world = vec4_neg(record->n_world);
 	record->obj = (t_object *)obj;
